Move components and pages around.
