+++ src/index.html
@@ -21,11 +21,11 @@
     <link
       rel="preload"
       as="style"
-      href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400;500&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@300;400;500&display=swap"
     />
     <link
       rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400;500&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@300;400;500&display=swap"
       media="print"
       onload="this.media='all'"
     />
@@ -89,6 +89,9 @@
         ></path>
       </svg>
     </a>
+
+    <!-- 🐉 Parcel needs to be able to see class names for Tailwind overrides to persist. -->
+    <div class="now-quote"></div>
 
     <script type="module" src="index.js"></script>
   </body>
