+++ src/pages/content/about.mdx
@@ -1,18 +1,17 @@
-import Emoji from "../components/emoji";
+import Emoji from "../../components/emoji";
 
-- Product engineer at [incident.io](https://incident.io) <Emoji emoji="🔥" /> in
-  London, previously [BBC News](https://www.bbc.co.uk/news) and
-  [Vidsy](https://vidsy.co)
-- [Projects](https://github.com/revett) on GitHub (@revett)
+- Open-source [projects](https://github.com/revett) on GitHub (@revett)
+- I started my career at [BBC News](https://www.bbc.co.uk/news) and was the CTO
+  of [Vidsy](https://vidsy.co)
 - Recent [adventures](https://instagram.com/cdrev) on Instagram (@cdrev)
 - B&W [photos](https://instagram.com/chadexbw) I've recently taken (@chadexbw)
 - Sport climbing and bouldering
   [logbook](https://www.8a.nu/user/charlie-revett/bouldering) (8a.nu)
-- [Twitter](https://twitter.com/revcd) (@revcd); not used a lot
+- Ramblings on [Twitter](https://twitter.com/revcd) (@revcd)
 - Natural language processing
   [academic papers](https://github.com/revett/academic-papers)
 - [LinkedIn](https://www.linkedin.com/in/cdrev) (@cdrev)
 - Coding [katas](https://www.codewars.com/users/revett) (@revett)
 
-Feel free to email me using [revcd at revdex.fastmail.com](#); this site is on
-[GitHub](https://github.com/revett/revcd.com)
+If you would like to get in touch, please
+[drop me an email](mailto:revcd-hello@revdex.fastmail.com)
