+++ src/pages/content/now.mdx
@@ -1,6 +1,6 @@
-import Emoji from "../components/emoji";
+import Emoji from "../../components/emoji";
 
-\> _Covering Summer / Autumn 2022_
+<p className="now-quote">Covering Summer / Autumn 2022</p>
 
 - Completing the renovation of my apartment in Bermondsey
 - Catching up on travelling: Brooklyn <Emoji emoji="🇺🇸" />,
