+++ src/pages/content/projects.mdx
